+++ src/Traits/FacadeTestCaseTrait.php
@@ -83,12 +83,6 @@ trait FacadeTestCaseTrait
 
     public function testFacadeRoot()
     {
-        if (defined('HHVM_VERSION')) {
-            if (version_compare(HHVM_VERSION, '3.1.0') === -1) {
-                return $this->markTestSkipped('Reflection API is broken on this version of HHVM.');
-            }
-        }
-
         $root = $this->getFacadeRoot();
         $class = $this->getFacadeClass();
         $reflection = new ReflectionClass($class);